backend/requirements.txt
flask
flask-cors
pdfminer.six
python-docx
flask-jwt-extended
flask-sqlalchemy